Ajout de la somme de contrôle notamment

+++ archive.sh
@@ -3,10 +3,27 @@
 # Importer le fichier conf
 . "./archive.conf"
 
-echo "$archiveURL"
+## Fonction d'écriture de log
+# $1 : emplacement des logs
+# $2 : checksum du fichier concerné
+# $3 : succès (0) / échec (1) de l'opération
+# $4 : si échec, description
+function ecrireLog() {
+	if [[ $3 -eq 0 ]]; then
+		echo "[ $(date +'%T - %d %b %Y') ] : Succès, checksum=$2\n" >> "$1"
+	else
+		echo "[ $(date +'%T - %d %b %Y') ] Échec, $4" >> "$1"
+	fi
+}
+
+## Fonction d'envoi de mail
+function envoyerMail() {
+	# TODO
+}
 
 # Vérifier si l'URL existe
 if wget -q --spider "$archiveURL"; then
+
 	# Télécharger l'archive
 	wget -q "$archiveURL"
 
@@ -16,23 +33,36 @@ if wget -q --spider "$archiveURL"; then
 	# Décompresser l'archive
 	unzip -q "$archive"
 
-	# sqldump
-	currentSql=$()
-	# Stocker la somme précédente dans un fichier externe caché
+	# Stocker la somme de contrôle du nouveau dump
+	currentChecksum=$(sha256sum ./*.sql | cut -d ' ' -f1)
 
-	# Archiver le nouveau dump
-	# TODO être plus spécifique sur le nom
-	tar -czf "$(date +'%Y%d%m')".tgz ./*.sql
+	# Vérifier les cahngements
+	if [[ "$(head -n 1 .prevChecksum)" != "$currentChecksum" ]]; then
 
-	# Supprimer les fichiers inutiles
-	rm -vf ./*.sql ./*.zip
+		# Archiver le nouveau dump
+		# TODO être plus spécifique sur le nom
+		tar -czf "$(date +'%Y%d%m')".tgz ./*.sql
 
-	# Pousser sur le serveur, avec SSH ou SFTP par ex
-	#scp fichier destination
+		# Supprimer les fichiers inutiles
+		rm -vf ./*.sql ./*.zip
 
-	# Logs
+		# Pousser sur le serveur, avec SSH ou SFTP par ex
+		#scp fichier destination
 
-	# Mail
+		# Si tout va bien, actualiser la somme de contrôle
+		echo "$currentChecksum" > .prevChecksum
+
+		# Logs
+
+		# Mail
+	else
+		echo "Le dump a la même somme de contrôle que précédemment."
+		# Mail
+	fi
+
 else
+
 	echo "L'URL de l'archive renseignée n'existe pas."
+
 fi
+
